Django-Poll-App/requirements/base.txt
dj-database-url==0.5.0
Django==3.2.4
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
asgiref==3.3.1
sqlparse==0.4.1

